requirements.txt
boto3 
pillow 
flask
python-dotenv